ai/requirements.txt
torch>=2.1.0
torchvision
efficientnet-pytorch>=0.7.1
timm>=0.9.0
grad-cam>=1.4.8
spacy>=3.7.0
fastapi>=0.110.0
uvicorn
grpcio>=1.62.0
grpcio-tools
Pillow
requests
python-dotenv
tqdm
matplotlib
scikit-learn